+++ packages/react/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @layerhub-io/react
 
+## 0.3.0
+
+### Minor Changes
+
+- Add background image support
+
+### Patch Changes
+
+- Updated dependencies
+  - @layerhub-io/core@0.3.0
+
 ## 0.2.6
 
 ### Patch Changes